import bugged, fixed compilation issues

+++ FbxDataExtractor/Util.cpp
@@ -3,8 +3,9 @@
 
 void Util::FbxVectorToArray(const FbxVector4& sourceVector, array<float>^ targetArray)
 {
-	for (size_t i = 0; i < targetArray->Length; i++)
+	for (int i = 0; i < targetArray->Length; i++)
 	{
 		targetArray[i] = sourceVector.mData[i];
 	}
 }
+
